backend/requirements.txt
flask==3.0.3
flask_cors
google.generativeai
python-dotenv
